refactor: extract analyseCommands, simplify PostgresEventStore
- Move command analysis to analyseCommands.ts
- Remove flattenEvents and posToNumber helpers (use ensureIsArray
  and inline parseInt directly)
- Extract withTransaction() to deduplicate COPY/batch transaction
  handling
- Simplify append() routing
