+++ packages/event-store-postgres/src/eventStore/PostgresEventStore.ts
@@ -1,4 +1,4 @@
-import { Pool, QueryResult } from "pg"
+import { Pool, PoolClient, QueryResult } from "pg"
 import {
     EventStore,
     DcbEvent,
@@ -15,69 +15,15 @@ import { dbEventConverter } from "./utils"
 import { readSqlWithCursor } from "./readSql"
 import { ensureInstalled } from "./ensureInstalled"
 import { LockStrategy, advisoryLocks } from "./lockStrategy"
-import { copyEventsToTable, copyConditionsToTempTable, ConditionRow } from "./copyWriter"
+import { copyEventsToTable, copyConditionsToTempTable } from "./copyWriter"
 import { getHighWaterMark, getLastPosition, checkBatchConditions, isConditionViolated } from "./queries"
+import { analyseCommands } from "./analyseCommands"
 
 const VALID_IDENTIFIER = /^[a-z_][a-z0-9_]{0,62}$/i
 const READ_BATCH_SIZE = 5000
 const COPY_THRESHOLD = 10
 const TAG_DELIMITER = "\x1F"
 const CONDITION_VIOLATED_SIGNAL = "APPEND_CONDITION_VIOLATED"
-
-function flattenEvents(cmd: AppendCommand): DcbEvent[] {
-    return ensureIsArray(cmd.events)
-}
-
-function posToNumber(pos: SequencePosition): number {
-    return parseInt(pos.toString())
-}
-
-/** Single pass over commands — extracts lock keys, conditions, and total event count. */
-function analyzeCommands(
-    commands: AppendCommand[],
-    lockStrategy: LockStrategy
-): {
-    totalEvents: number
-    lockKeys: bigint[]
-    conditions: ConditionRow[]
-    eventIterator: () => Iterable<DcbEvent>
-} {
-    const allKeys = new Set<bigint>()
-    const conditions: ConditionRow[] = []
-    let totalEvents = 0
-
-    for (let i = 0; i < commands.length; i++) {
-        const cmd = commands[i]
-        const evts = flattenEvents(cmd)
-        totalEvents += evts.length
-
-        for (const k of lockStrategy.computeKeys(evts, cmd.condition)) {
-            allKeys.add(k)
-        }
-
-        if (cmd.condition?.after) {
-            for (const item of cmd.condition.failIfEventsMatch.items) {
-                conditions.push({
-                    cmdIdx: i,
-                    types: item.types ?? [],
-                    tags: item.tags?.values ?? [],
-                    afterPos: posToNumber(cmd.condition.after)
-                })
-            }
-        }
-    }
-
-    return {
-        totalEvents,
-        lockKeys: [...allKeys],
-        conditions,
-        eventIterator: function* () {
-            for (const cmd of commands) {
-                for (const evt of flattenEvents(cmd)) yield evt
-            }
-        }
-    }
-}
 
 export interface PostgresEventStoreOptions {
     pool: Pool
@@ -111,13 +57,12 @@ export class PostgresEventStore implements EventStore {
         const commands = ensureIsArray(command)
 
         if (commands.length === 1) {
-            const evts = flattenEvents(commands[0])
+            const evts = ensureIsArray(commands[0].events)
             if (evts.length === 0) throw new Error("Cannot append zero events")
 
-            if (evts.length <= this.copyThreshold) {
-                return this.appendViaFunction(evts, commands[0].condition)
-            }
-            return this.appendViaCopy(evts, commands[0].condition)
+            return evts.length <= this.copyThreshold
+                ? this.appendViaFunction(evts, commands[0].condition)
+                : this.appendViaCopy(evts, commands[0].condition)
         }
 
         return this.appendBatch(commands)
@@ -132,9 +77,7 @@ export class PostgresEventStore implements EventStore {
 
             let result: QueryResult
             while ((result = await client.query(`FETCH ${READ_BATCH_SIZE} FROM ${cursorName}`))?.rows?.length) {
-                for (const ev of result.rows) {
-                    yield dbEventConverter.fromDb(ev)
-                }
+                for (const ev of result.rows) yield dbEventConverter.fromDb(ev)
             }
         } finally {
             await client.query("ROLLBACK").catch(() => {})
@@ -142,11 +85,9 @@ export class PostgresEventStore implements EventStore {
         }
     }
 
-    // ─── Stored procedure path (1 round-trip for small appends) ──
-
+    /** Stored procedure — single round-trip for small appends (≤ copyThreshold events). */
     private async appendViaFunction(evts: DcbEvent[], condition?: AppendCondition): Promise<SequencePosition> {
         const lockKeys = this.lockStrategy.computeKeys(evts, condition)
-
         const types: string[] = []
         const tags: string[] = []
         const payloads: string[] = []
@@ -157,13 +98,17 @@ export class PostgresEventStore implements EventStore {
             payloads.push(serializePayload(evt))
         }
 
-        const conditionItems = condition ? serializeConditionItems(condition) : null
-        const afterPos = condition?.after ? posToNumber(condition.after) : null
-
         try {
             const result = await this.pool.query(
                 `SELECT ${this.appendFunctionName}($1::text[], $2::text[], $3::text[], $4::bigint[], $5::jsonb, $6::bigint) as pos`,
-                [types, tags, payloads, lockKeys, conditionItems, afterPos]
+                [
+                    types,
+                    tags,
+                    payloads,
+                    lockKeys,
+                    condition ? serializeConditionItems(condition) : null,
+                    condition?.after ? parseInt(condition.after.toString()) : null
+                ]
             )
             return SequencePosition.fromString(String(result.rows[0].pos))
         } catch (err) {
@@ -174,73 +119,51 @@ export class PostgresEventStore implements EventStore {
         }
     }
 
-    // ─── COPY path (single command, large event count) ───────────
-
+    /** COPY FROM STDIN — high throughput for large single-command appends. */
     private async appendViaCopy(evts: DcbEvent[], condition?: AppendCondition): Promise<SequencePosition> {
         const lockKeys = this.lockStrategy.computeKeys(evts, condition)
+        return this.withTransaction(async client => {
+            if (lockKeys.length > 0) await this.lockStrategy.acquire(client, lockKeys, this.tableName)
 
-        const client = await this.pool.connect()
-        try {
-            await client.query("BEGIN TRANSACTION ISOLATION LEVEL READ COMMITTED")
-
-            if (lockKeys.length > 0) {
-                await this.lockStrategy.acquire(client, lockKeys, this.tableName)
-            }
-
-            if (condition) {
-                if (await isConditionViolated(client, this.tableName, condition)) {
-                    await client.query("ROLLBACK")
-                    throw new AppendConditionError(condition)
-                }
+            if (condition && (await isConditionViolated(client, this.tableName, condition))) {
+                throw new AppendConditionError(condition)
             }
 
             await copyEventsToTable(client, this.tableName, evts)
-
-            const pos = await getLastPosition(client, this.tableName)
-            await client.query("COMMIT")
-            return SequencePosition.fromString(String(pos))
-        } catch (err) {
-            if (err instanceof AppendConditionError) throw err
-            await client.query("ROLLBACK").catch(() => {})
-            throw err
-        } finally {
-            client.release()
-        }
+            return SequencePosition.fromString(String(await getLastPosition(client, this.tableName)))
+        })
     }
 
-    // ─── Batch path (multiple commands, COPY + temp table) ───────
-
+    /** Batch COPY + temp table — multiple commands with per-command condition checking. */
     private async appendBatch(commands: AppendCommand[]): Promise<SequencePosition> {
-        const { totalEvents, lockKeys, conditions, eventIterator } = analyzeCommands(commands, this.lockStrategy)
+        const { totalEvents, lockKeys, conditions, eventIterator } = analyseCommands(commands, this.lockStrategy)
         if (totalEvents === 0) throw new Error("Cannot append zero events")
 
-        const client = await this.pool.connect()
-        try {
-            await client.query("BEGIN TRANSACTION ISOLATION LEVEL READ COMMITTED")
-
-            if (lockKeys.length > 0) {
-                await this.lockStrategy.acquire(client, lockKeys, this.tableName)
-            }
+        return this.withTransaction(async client => {
+            if (lockKeys.length > 0) await this.lockStrategy.acquire(client, lockKeys, this.tableName)
 
             const highWaterMark = await getHighWaterMark(client, this.tableName)
-
             await copyEventsToTable(client, this.tableName, eventIterator())
 
             if (conditions.length > 0) {
                 await copyConditionsToTempTable(client, conditions)
-
                 const failedIdx = await checkBatchConditions(client, this.tableName, highWaterMark)
-                if (failedIdx !== null) {
-                    await client.query("ROLLBACK")
-                    throw new AppendConditionError(commands[failedIdx].condition!, failedIdx)
-                }
+                if (failedIdx !== null) throw new AppendConditionError(commands[failedIdx].condition!, failedIdx)
             }
 
-            const pos = await getLastPosition(client, this.tableName)
+            return SequencePosition.fromString(String(await getLastPosition(client, this.tableName)))
+        })
+    }
+
+    /** Run a callback inside a READ COMMITTED transaction, with rollback on error. */
+    private async withTransaction<T>(fn: (client: PoolClient) => Promise<T>): Promise<T> {
+        const client = await this.pool.connect()
+        try {
+            await client.query("BEGIN TRANSACTION ISOLATION LEVEL READ COMMITTED")
+            const result = await fn(client)
             await client.query("COMMIT")
-            return SequencePosition.fromString(String(pos))
+            return result
         } catch (err) {
-            if (err instanceof AppendConditionError) throw err
             await client.query("ROLLBACK").catch(() => {})
             throw err
         } finally {
